backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.28.1
